+++ apps/projects/app/components/App/index.js
@@ -1,4 +1,3 @@
 export { default as ErrorBoundary } from './ErrorBoundary'
 export { default as NewIssueButton } from './NewIssueButton'
 export { default as NewProjectButton } from './NewProjectButton'
-export { default as AppTitleButton } from './AppTitleButton'
