+++ apps/projects/package.json
@@ -32,7 +32,7 @@
     "@aragon/api": "1.1.0",
     "@aragon/api-react": "1.0.0-beta.2",
     "@aragon/apps-vault": "^4.1.0",
-    "@aragon/ui": "0.33.0",
+    "@aragon/ui": "1.0.0-alpha.19",
     "@babel/polyfill": "^7.2.5",
     "@tps/ui": "^0.0.1",
     "apollo-boost": "^0.1.22",
